requirements.txt
biopython==1.79
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
freesasa==2.1.0
idna==3.3
isort==5.10.1
lxml==4.8.0
mypy-extensions==0.4.3
numpy
pandas
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
